#87 - refactor: 팔로우 Controller 단일 책임 원칙 적용
